pyproject.toml
[project]
name = "kgfproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "fastapi[standard]>=0.135.1",
    "pydantic-settings>=2.13.1",
    "sqladmin[full]>=0.23.0",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "black>=26.3.0",
    "mypy>=1.19.1",
    "ruff>=0.15.5",
]

[tool.mypy]
strict = true
